Delete extra file for push notification registration
