+++ Decred Wallet/UIApplication.swift
@@ -4,6 +4,7 @@
 //  Copyright © 2018 The Decred developers.
 //  see LICENSE for details.
 
+import UserNotifications
 import SlideMenuControllerSwift
 
 extension UIApplication {
@@ -25,5 +26,33 @@ extension UIApplication {
             return topViewController(slide.mainViewController)
         }
         return viewController
+    }
+    
+    func registerForPushNotifications() {
+        if #available(iOS 10.0, *) {
+            let center = UNUserNotificationCenter.current()
+            center.requestAuthorization(
+            options: [.badge, .alert, .sound]) { [weak self] granted, _ in
+                DispatchQueue.main.async {
+                    if granted {
+                        self?.registerForRemoteNotifications()
+                    }
+                }
+            }
+        } else {
+            let type: UIUserNotificationType = [
+                UIUserNotificationType.badge,
+                UIUserNotificationType.alert,
+                UIUserNotificationType.sound,
+                ]
+            
+            let setting = UIUserNotificationSettings(
+                types: type,
+                categories: nil
+            )
+            
+            UIApplication.shared.registerUserNotificationSettings(setting)
+            UIApplication.shared.registerForRemoteNotifications()
+        }
     }
 }
